Update header and rename pages for dibbs site
